ADT.Maybe

Maybe encapulates the idea of a value that might not be there. This is often a failed computation, but in some cases an empty value may be the expected behaviour.

A %Maybe{} value can either be Just some_value or Nothing. More typical of Elixir is to use {:ok, some_value}, or {:error, some_reason}. By contrast, %Maybe{} has an implied :ok, if there is a value in the maybe key.

Please note that this approach does not track error reasons, as a Nothing value may not be an error. If you are looking for error tracking behaviour, consider %MaybePlus{}

Functions

from_status_tuple(arg)

Common useage

just?(maybe)

nothing?(x)