Witchcraft.Applicative protocol

Applicative functors provide a method of applying a function contained in a datatype to a value of the same datatype.

Types

t :: term

Functions

apply(data, function)

pure(member)

Specs

pure(any) :: any