Witchcraft v0.4.1 Witchcraft.Functor.Operator

Functions

func <~ args

Specs

(any -> any) <~ any :: any

Alias for lift. As we’ll see with Witchcraft.Applicative, this arrow points in the direction of data flow (just like |>), but we often prefer the function on the left side.

iex> (&(&1 * 10)) <~ [1,2,3]
[10, 20, 30]

args ~> func

Specs

any ~> (any -> any) :: any

Alias for lift and <~, but with data flowing to the right.

iex> [1,2,3] ~> &(&1 * 10)
[10, 20, 30]