Witchcraft v0.4.0 Witchcraft.Functor.Property

Check samples of your functor to confirm that your data adheres to the functor properties. All members of your datatype should adhere to these rules. They are placed here as a quick way to spotcheck some of your values.

Functions

spotcheck(context, f, g, typecheck)

Specs

spotcheck(any, (any -> any), (any -> any), (any -> boolean)) :: boolean

Spotcheck all functor properties

iex> alias Algae.Id, as: Id
iex> defmodule Typecheck do
...>   def is_id(%Id{id: _}), do: true
...>   def is_id(_), do: false
...> end
iex> spotcheck(%Id{id: 42}, &(&1 + 1), &(&1 * 2), &Typecheck.is_id/1)
true

spotcheck_associates_object(context, func, typecheck)

Specs

spotcheck_associates_object(any, (any -> any), (any -> boolean)) :: boolean

Check that lifting a function into some context returns a member of the target type

iex> alias Algae.Id, as: Id
iex> defmodule Typecheck do
...>   def is_id(%Id{id: _}), do: true
...>   def is_id(_), do: false
...> end
iex> spotcheck_associates_object(%Id{id: 42}, &Quark.id/1, &Typecheck.is_id/1)
true

spotcheck_preserve_compositon(context, f, g)

Specs

spotcheck_preserve_compositon(any, (any -> any), (any -> any)) :: boolean

Check that lifting a composed function is the same as lifting functions in sequence

iex> spotcheck_preserve_compositon(%Algae.Id{id: 5}, &(&1 + 1), &(&1 * 10))
true

spotcheck_preserve_identity(context, func)

Specs

spotcheck_preserve_identity(any, (any -> any)) :: boolean

Check that lifting a function does not interfere with identity. In other words, lifting id(a) shoud be the same as the identity of lifting a.

 A ---- id ----> A

 |               |
(f)             (f)
 |               |
 v               v

 B ---- id ----> B

iex> spotcheck_preserve_identity(%Algae.Id{id: 7}, &(&1 + 1))
true