Witchcraft.Utility

Functions

compose(func_list)

Specs

compose([(... -> any)]) :: (... -> any)

Function composition, from the back of the lift to the front

Example

iex> sum_plus_one = Witchcraft.Utility.compose([&(&1 + 1), &(Enum.sum(&1))]) iex> [1,2,3] |> sum_plus_one.() 7

constant(, b)

Specs

constant(any, any) :: any

Alias for second/2

first(a, )

Specs

first(any, any) :: any

Return the first of two arguments. Can be used to repeatedly apply the same value in functions such as folds.

Examples

iex> Witchcraft.Utility.first(43, 42)
43

iex> Enum.reduce([1,2,3], [42], &Witchcraft.Utility.first(&1, &2))
3

id(a)

Specs

id(any) :: any

Do nothing to an argument; just return it

Examples

iex> Witchcraft.Utility.id("88 miles per hour")
"88 miles per hour"

iex> Witchcraft.Utility.id(42)
42

iex> Enum.map([1,2,3], &Witchcraft.Utility.id&1)
[1,2,3]

reverse_compose(func_list)

Specs

reverse_compose([(... -> any)]) :: (... -> any)

Compose functions, from the head of the list of functions. The is the reverse order versus what one would normally expect (left to right rather than right to left).

Example

iex> sum_plus_one = Witchcraft.Utility.reverse_compose([&(Enum.sum(&1)), &(&1 + 1)]) iex> [1,2,3] |> sum_plus_one.() 7

second(, b)

Specs

second(any, any) :: any

Return the second of two arguments. Can be used to repeatedly apply the same value in functions such as folds.

Examples

iex> Witchcraft.Utility.second(43, 42)
42

iex> Enum.reduce([1,2,3], [], &Witchcraft.Utility.second(&1, &2))
[]